Fix configuration file mechanism.
First, there may be a file 'config.pl' elsewherewhere in @INC, so let's
make it less ambiguous.  Also, later versions of perl don't include "."
in @INC, so it may happen that the file exists in ".", is readable in ".",
but can't be read in using do(...).  In particular, the removal of "."
from @INC seems to be a change inserted into security releases
of the perl package, version 5.24.1*, in Debian 9.2.